requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
oauth2==1.9.0.post1
oauthlib==3.1.0
panda3d==1.10.3
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
sklearn==0.0
urllib3==1.25.8
Werkzeug==1.0.0
